requirements.txt
flask
joblib
pandas
scikit-learn
requests
beautifulsoup4
openai
python-dotenv
gunicorn
